Remove old version of EntitiesContainer component
Move new version into the path of the old one.
