+++ gsa/CMakeLists.txt
@@ -340,7 +340,6 @@ set (GSA_JS_SRC_FILES
      ${GSA_SRC_DIR}/src/web/components/table/table.js
      ${GSA_SRC_DIR}/src/web/entities/actions.js
      ${GSA_SRC_DIR}/src/web/entities/container.js
-     ${GSA_SRC_DIR}/src/web/entities/container2.js
      ${GSA_SRC_DIR}/src/web/entities/entitynametabledata.js
      ${GSA_SRC_DIR}/src/web/entities/footer.js
      ${GSA_SRC_DIR}/src/web/entities/header.js
